clean up elements
